extracting a helper
